Minor cleanup and readability refactoring
