+++ partiql-eval/src/eval/expr/mod.rs
@@ -1,5 +1,3 @@
-pub(crate) mod eval_wrapper;
-
 mod base_table;
 pub(crate) use base_table::*;
 mod coll;
