+++ partiql-eval/src/eval/expr/pattern_match.rs
@@ -1,6 +1,6 @@
 use crate::error::PlanningError;
 
-use crate::eval::expr::eval_wrapper::{TernaryValueExpr, UnaryValueExpr};
+use crate::eval::eval_expr_wrapper::{TernaryValueExpr, UnaryValueExpr};
 use crate::eval::expr::{BindError, BindEvalExpr, EvalExpr};
 use partiql_types::TYPE_STRING;
 use partiql_value::Value;
